requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9
pandas
xlsxwriter==3.1.9
python-multipart==0.0.9
pydantic==2.6.1 